+++ mayan/apps/common/apps.py
@@ -146,10 +146,7 @@ class CommonApp(MayanAppConfig):
         )
 
         menu_about.bind_links(
-            links=(
-                link_tools, link_setup, link_about, link_knowledge_base,
-                link_book, link_store, link_support, link_license
-            )
+            links=(link_tools, link_setup, link_about, link_license)
         )
 
         menu_topbar.bind_links(