+++ HISTORY.rst
@@ -136,6 +136,7 @@
   - Move templates into sub-folders and shorten their names.
 
 - Fix double separator in the user menu.
+- Theme updates.
 
 4.5.9 (XXXX-XX-XX)
 ==================